interpolate replaces upsample - again 7
