+++ models/checkpointed_unet.py
@@ -1,4 +1,4 @@
-import torch
+import torch, utils
 # import torch should be first. Unclear issue, mentioned here: https://github.com/pytorch/pytorch/issues/2083
 from torch import nn
 from torch.utils.checkpoint import checkpoint_sequential
@@ -44,18 +44,6 @@ class EncodingBlock(nn.Module):
         modules = get_modules(self.encoding_block)
         return checkpoint_sequential(modules, segments, input_data)
 
-class Interpolate(nn.Module):
-    def __init__(self, mode, scale_factor):
-        super(Interpolate, self).__init__()
-        self.interp = nn.functional.interpolate
-        self.scale_factor = scale_factor
-        self.mode = mode
-
-    def forward(self, x):
-        x = self.interp(x, scale_factor=self.scale_factor, mode=self.mode, align_corners=False)
-        return x
-
-
 class DecodingBlock(nn.Module):
     """Module in the decoding section of the UNet"""
 
@@ -63,7 +51,7 @@ class DecodingBlock(nn.Module):
         super().__init__()
         up_modules = []
         if upsampling:
-            self.up = nn.Sequential(Interpolate(mode='bilinear', scale_factor=2),
+            self.up = nn.Sequential(utils.Interpolate(mode='bilinear', scale_factor=2),
                                     nn.Conv2d(in_size, out_size, kernel_size=1))
             self.upsampling = True
         else:
